mlrunnerserver/requirements.txt
apache-libcloud==3.2.0
asgiref==3.2.10
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
Django==3.1.1
idna==2.10
paramiko==2.7.2
pycparser==2.20
pycrypto==2.6.1
PyNaCl==1.4.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
